Update allCards.json - Set 9
2025-08-22 2.2.1
Format addition: Because of the two new rarities, the fields epicId and iconicId are added. They work the same as enchantedId
Format addition: Instead of creating the equivalent of nonEnchantedId for Epic- and Iconic-rarity cards, a new consolidated field has been added: baseId. This field points to the base card within the normal numbering from Promo cards and from cards with Epic, Enchanted, or Iconic rarity
Warning: Because the aforementioned baseId consolidates them, nonEnchantedId and nonPromoId will be removed in the next update on September 5th
Technical note: Previously the zipfiles were compressed with the LZMA-algorithm. From this release on, the DEFLATE-algorithm will be used. The zipfiles are slightly larger, but DEFLATE has larger compatibility. (If this doesn't mean anything to you, it probably doesn't affect you)
Corrected the date when sets 1 to 4 are rotated out of Core Constructed from "2025-09-19" to "2025-09-05", the date when Set 9 fully releases
Some foil and varnish types were renamed:
The 'Cold' foil type was renamed to 'Silver', and the 'Freeform' foil type was renamed to 'Freeform2'
varnishType of the D23 version of Mickey Mouse - Playful Sorcerer (ID 1432) was renamed from 'Gold' to 'RainbowHotFoil'
varnishType of the reward card from "Illumineer's Quest: Palace Heist" (ID 1935) was renamed from 'InkColoredHotFoil' to 'MatteHotFoil'
Added cardmarketId to externalLinks of the promo version of A Whole New World (ID 1199)